+++ app/models/palette.rb
@@ -1,0 +1,4 @@
+class Palette < ApplicationRecord
+  belongs_to :user, optional: true
+  has_many :colors, dependent: :destroy
+end