+++ db/migrate/20260306202136_create_palettes.rb
@@ -1,0 +1,10 @@
+class CreatePalettes < ActiveRecord::Migration[8.1]
+  def change
+    create_table :palettes do |t|
+      t.string :name
+      t.references :user, null: true, foreign_key: true
+
+      t.timestamps
+    end
+  end
+end