+++ db/migrate/20260306202138_create_colors.rb
@@ -1,0 +1,11 @@
+class CreateColors < ActiveRecord::Migration[8.1]
+  def change
+    create_table :colors do |t|
+      t.string :hex_code
+      t.integer :position
+      t.references :palette, null: false, foreign_key: true
+
+      t.timestamps
+    end
+  end
+end